# More efficient ColumnChunk string dictionary caching